chore: Move formula code to one dir, split function definitions into smaller files [PT-186305453]
