+++ v3/src/models/document/create-document-model.ts
@@ -3,7 +3,7 @@ import { ITileEnvironment } from "../tiles/tile-environment"
 import { DocumentModel, IDocumentModelSnapshot } from "./document"
 import { IDocumentEnvironment } from "./document-environment"
 import { SharedModelDocumentManager } from "./shared-model-document-manager"
-import { FormulaManager } from "../data/formula-manager"
+import { FormulaManager } from "../formula/formula-manager"
 import { ISharedDataSet, SharedDataSet, kSharedDataSetType } from "../shared/shared-data-set"
 
 /**
