+++ v3/src/models/tiles/tile-environment.ts
@@ -2,7 +2,7 @@ import { getEnv, IAnyStateTreeNode } from "mobx-state-tree"
 // There's nothing wrong in explicit ISharedModelManager interface, but probably dependency cycle could have been also
 // avoided using `import type { SharedModelManager } from ...`.
 import { ISharedModelManager } from "../shared/shared-model-manager"
-import type { FormulaManager } from "../data/formula-manager"
+import type { FormulaManager } from "../formula/formula-manager"
 
 export interface ITileEnvironment {
   sharedModelManager?: ISharedModelManager
